fix: alakarkinos recipe sync
