+++ gradle.properties
@@ -18,7 +18,7 @@ loader_version_range=[2,)
 mod_id=ars_nouveau
 mod_name=Ars Nouveau
 mod_license=LGPL
-mod_version=5.4.2
+mod_version=5.4.2-beta.1
 mod_group_id=com.hollingsworth.ars_nouveau
 mod_authors=Bailey Hollingsworth
 mod_description=Craft spells, create powerful baubles, and summon magical creatures!
